feat: moved endpoint apiUrlFromSwarmHost
