+++ src/constants/index.ts
@@ -1,7 +1,8 @@
 /* eslint-disable no-console */
 import moment from 'moment'
+import { API_URL } from '~/utils/apiUrlFromSwarmHost'
 
-const { origin, host } = window.location
+const { origin } = window.location
 
 moment.relativeTimeThreshold('h', 24)
 
@@ -21,35 +22,7 @@ export const isDevelopment = !!(
   origin === 'https://sphinx-jarvis-david.sphinx1.repl.co'
 )
 
-const getUrlFormEnv = () => import.meta.env.VITE_APP_API_URL
-
-export const API_URL = getUrlFormEnv() || apiUrlFromSwarmHost() || 'https://knowledge-graph.sphinx.chat'
-
 export const isChileGraph = API_URL.includes('boltwall')
-
-function apiUrlFromSwarmHost(): string | undefined {
-  // for swarm deployments, always point to "boltwall"
-  // for now, only if the URL contains "swarm"
-  const originUrl = window.location.origin
-
-  let url = originUrl
-
-  if (host.includes('swarm')) {
-    if (host.startsWith('nav')) {
-      const hostArray = host.split('.')
-
-      hostArray[0] = 'boltwall'
-
-      const finalHost = hostArray.join('.')
-
-      url = `https://${finalHost}`
-    }
-  } else if (originUrl === 'https://second-brain.sphinx.chat' || origin.includes('localhost')) {
-    url = 'https://knowledge-graph.sphinx.chat'
-  }
-
-  return `${url}/api`
-}
 
 export const AWS_IMAGE_BUCKET_URL = 'https://stakwork-uploads.s3.amazonaws.com/'
 export const CLOUDFRONT_IMAGE_BUCKET_URL = 'https://d1gd7b7slyku8k.cloudfront.net/'
