+++ src/hooks/useSockets/index.ts
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from 'react'
 import { Socket, io } from 'socket.io-client'
-import { API_URL } from '~/constants'
+import { API_URL } from '~/utils/apiUrlFromSwarmHost'
 
 const useSocket = (): Socket | null => {
   const [socket, setSocket] = useState<Socket | null>(null)
