+++ src/network/api/index.ts
@@ -1,5 +1,5 @@
-import { API_URL } from '~/constants'
 import { getSignedMessageFromRelay } from '~/utils'
+import { API_URL } from '~/utils/apiUrlFromSwarmHost'
 
 const request = async <Res>(url: string, config?: RequestInit): Promise<Res> => {
   const admin = localStorage.getItem('admin')
